src/api/py/pyproject.toml
[tool.poetry]
authors = ["Tom Brovender <Or.Shemtov@cellebrite.com>"]
description = ""
name = "api"
version = "0.1.0"

[tool.poetry.dependencies]
aws-lambda-typing = "^2.17.1"
boto3 = "^1.26.162"
python = "^3.10"
pyyaml = "^6.0.1"

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core"]
